Update dive log screens and defaults
