+++ src/Components/SelectWGradientborderForPrivacy.tsx
@@ -138,7 +138,7 @@ const SelectWGradientBorderForPrivacy: FunctionComponent<ComponentProps> = ({
 
 const styles = StyleSheet.create({
   container: {
-    marginTop: 20,
+    marginTop: 10,
     flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'space-between',
