+++ src/Components/ui/RatingsInputComp.tsx
@@ -43,7 +43,7 @@ const RatingsInputComp: FunctionComponent<ComponentProps> = ({
 
 const styles = StyleSheet.create({
   ratingContainer: {
-    marginTop: 20,
+    marginTop: 10,
     flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'space-between',
